+++ MatrixSDK/Crypto/MXCrypto.m
@@ -36,7 +36,7 @@
 #import "MXOutgoingRoomKeyRequestManager.h"
 #import "MXIncomingRoomKeyRequestManager.h"
 
-#import "MXDeviceVerificationManager_Private.h"
+#import "MXKeyVerificationManager_Private.h"
 #import "MXDeviceInfo_Private.h"
 #import "MXCrossSigning_Private.h"
 
@@ -1550,7 +1550,7 @@ NSTimeInterval kMXCryptoMinForceSessionPeriod = 3600.0; // one hour
 
         incomingRoomKeyRequestManager = [[MXIncomingRoomKeyRequestManager alloc] initWithCrypto:self];
 
-        _deviceVerificationManager = [[MXDeviceVerificationManager alloc] initWithCrypto:self];
+        _keyVerificationManager = [[MXKeyVerificationManager alloc] initWithCrypto:self];
 
         _crossSigning = [[MXCrossSigning alloc] initWithCrypto:self];