+++ MatrixSDK/Crypto/Verification/MXKeyVericationManager.m
@@ -15,8 +15,8 @@
  limitations under the License.
  */
 
-#import "MXDeviceVerificationManager.h"
-#import "MXDeviceVerificationManager_Private.h"
+#import "MXKeyVerificationManager.h"
+#import "MXKeyVerificationManager_Private.h"
 
 #import "MXSession.h"
 #import "MXCrypto_Private.h"
@@ -33,11 +33,11 @@
 
 #pragma mark - Constants
 
-NSString *const MXDeviceVerificationErrorDomain = @"org.matrix.sdk.verification";
-NSString *const MXDeviceVerificationManagerNewRequestNotification       = @"MXDeviceVerificationManagerNewRequestNotification";
-NSString *const MXDeviceVerificationManagerNotificationRequestKey       = @"MXDeviceVerificationManagerNotificationRequestKey";
-NSString *const MXDeviceVerificationManagerNewTransactionNotification   = @"MXDeviceVerificationManagerNewTransactionNotification";
-NSString *const MXDeviceVerificationManagerNotificationTransactionKey   = @"MXDeviceVerificationManagerNotificationTransactionKey";
+NSString *const MXKeyVerificationErrorDomain = @"org.matrix.sdk.verification";
+NSString *const MXKeyVerificationManagerNewRequestNotification       = @"MXKeyVerificationManagerNewRequestNotification";
+NSString *const MXKeyVerificationManagerNotificationRequestKey       = @"MXKeyVerificationManagerNotificationRequestKey";
+NSString *const MXKeyVerificationManagerNewTransactionNotification   = @"MXKeyVerificationManagerNewTransactionNotification";
+NSString *const MXKeyVerificationManagerNotificationTransactionKey   = @"MXKeyVerificationManagerNotificationTransactionKey";
 
 // Transaction timeout in seconds
 NSTimeInterval const MXTransactionTimeout = 10 * 60.0;
@@ -45,16 +45,16 @@ NSTimeInterval const MXTransactionTimeout = 10 * 60.0;
 // Request timeout in seconds
 NSTimeInterval const MXRequestDefaultTimeout = 5 * 60.0;
 
-static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
+static NSArray<MXEventTypeString> *kMXKeyVerificationManagerDMEventTypes;
 
 
-@interface MXDeviceVerificationManager ()
+@interface MXKeyVerificationManager ()
 {
     // The queue to run background tasks
     dispatch_queue_t cryptoQueue;
 
     // All running transactions
-    MXUsersDevicesMap<MXDeviceVerificationTransaction*> *transactions;
+    MXUsersDevicesMap<MXKeyVerificationTransaction*> *transactions;
     // Timer to cancel transactions
     NSTimer *transactionTimeoutTimer;
 
@@ -69,7 +69,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 }
 @end
 
-@implementation MXDeviceVerificationManager
+@implementation MXKeyVerificationManager
 
 #pragma mark - Public methods -
 
@@ -157,7 +157,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 - (void)beginKeyVerificationWithUserId:(NSString*)userId
                            andDeviceId:(NSString*)deviceId
                                 method:(NSString*)method
-                               success:(void(^)(MXDeviceVerificationTransaction *transaction))success
+                               success:(void(^)(MXKeyVerificationTransaction *transaction))success
                                failure:(void(^)(NSError *error))failure
 {
     [self beginKeyVerificationWithUserId:userId andDeviceId:deviceId dmRoomId:nil dmEventId:nil method:method success:success failure:failure];
@@ -168,7 +168,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
                               dmRoomId:(nullable NSString*)dmRoomId
                              dmEventId:(nullable NSString*)dmEventId
                                 method:(NSString*)method
-                               success:(void(^)(MXDeviceVerificationTransaction *transaction))success
+                               success:(void(^)(MXKeyVerificationTransaction *transaction))success
                                failure:(void(^)(NSError *error))failure
 {
     NSLog(@"[MXKeyVerification] beginKeyVerification: device: %@:%@ roomId: %@ method:%@", userId, deviceId, dmRoomId, method);
@@ -176,7 +176,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     // Make sure we have other device keys
     [self loadDeviceWithDeviceId:deviceId andUserId:userId success:^(MXDeviceInfo *otherDevice) {
 
-        MXDeviceVerificationTransaction *transaction;
+        MXKeyVerificationTransaction *transaction;
         NSError *error;
 
         // We support only SAS at the moment
@@ -197,8 +197,8 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
         }
         else
         {
-            error = [NSError errorWithDomain:MXDeviceVerificationErrorDomain
-                                        code:MXDeviceVerificationUnsupportedMethodCode
+            error = [NSError errorWithDomain:MXKeyVerificationErrorDomain
+                                        code:MXKeyVerificationUnsupportedMethodCode
                                     userInfo:@{
                                                NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unsupported verification method: %@", method]
                                                }];
@@ -221,13 +221,13 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     }];
 }
 
-- (void)transactions:(void(^)(NSArray<MXDeviceVerificationTransaction*> *transactions))complete
+- (void)transactions:(void(^)(NSArray<MXKeyVerificationTransaction*> *transactions))complete
 {
     MXWeakify(self);
     dispatch_async(self->cryptoQueue, ^{
         MXStrongifyAndReturnIfNil(self);
 
-        NSArray<MXDeviceVerificationTransaction*> *transactions = self->transactions.allObjects;
+        NSArray<MXKeyVerificationTransaction*> *transactions = self->transactions.allObjects;
         dispatch_async(dispatch_get_main_queue(), ^{
             complete(transactions);
         });
@@ -252,8 +252,8 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     }
     else
     {
-        NSError *error = [NSError errorWithDomain:MXDeviceVerificationErrorDomain
-                                             code:MXDeviceVerificationUnknownIdentifier
+        NSError *error = [NSError errorWithDomain:MXKeyVerificationErrorDomain
+                                             code:MXKeyVerificationUnknownIdentifier
                                          userInfo:@{
                                                     NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unknown id or not supported transport"]
                                                     }];
@@ -297,7 +297,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     MXKeyVerification *keyVerification;
 
     // First, check if this is a transaction in progress
-    MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:keyVerificationId];
+    MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:keyVerificationId];
     if (transaction)
     {
         keyVerification = [MXKeyVerification new];
@@ -325,7 +325,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 {
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
-        kMXDeviceVerificationManagerDMEventTypes = @[
+        kMXKeyVerificationManagerDMEventTypes = @[
                                                     kMXEventTypeStringKeyVerificationStart,
                                                     kMXEventTypeStringKeyVerificationAccept,
                                                     kMXEventTypeStringKeyVerificationKey,
@@ -375,7 +375,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
 - (void)acceptVerificationRequest:(MXKeyVerificationRequest*)request
                            method:(NSString*)method
-                          success:(void(^)(MXDeviceVerificationTransaction *transaction))success
+                          success:(void(^)(MXKeyVerificationTransaction *transaction))success
                           failure:(void(^)(NSError *error))failure
 {
     NSLog(@"[MXKeyVerification] acceptVerificationByDMRequest: event: %@", request.requestId);
@@ -384,8 +384,8 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     NSString *fromDevice = request.fromDevice;
     if (!fromDevice)
     {
-        NSError *error = [NSError errorWithDomain:MXDeviceVerificationErrorDomain
-                                             code:MXDeviceVerificationUnknownDeviceCode
+        NSError *error = [NSError errorWithDomain:MXKeyVerificationErrorDomain
+                                             code:MXKeyVerificationUnknownDeviceCode
                                          userInfo:@{
                                                     NSLocalizedDescriptionKey: [NSString stringWithFormat:@"from_device not found"]
                                                     }];
@@ -412,7 +412,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     MXTransactionCancelCode *cancelCode = MXTransactionCancelCode.user;
 
     // If there is transaction in progress, cancel it
-    MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:request.requestId];
+    MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:request.requestId];
     if (transaction)
     {
         [self cancelTransaction:transaction code:cancelCode];
@@ -442,7 +442,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
 #pragma mark - Transactions
 
-- (MXHTTPOperation*)sendToOtherInTransaction:(MXDeviceVerificationTransaction*)transaction
+- (MXHTTPOperation*)sendToOtherInTransaction:(MXKeyVerificationTransaction*)transaction
                                    eventType:(NSString*)eventType
                                      content:(NSDictionary*)content
                                      success:(void (^)(void))success
@@ -470,7 +470,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 }
 
 
-- (void)cancelTransaction:(MXDeviceVerificationTransaction*)transaction code:(MXTransactionCancelCode*)code
+- (void)cancelTransaction:(MXKeyVerificationTransaction*)transaction code:(MXTransactionCancelCode*)code
 {
     NSLog(@"[MXKeyVerification] cancelTransaction. code: %@", code.value);
     
@@ -597,7 +597,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     // Make sure we have other device keys
     [self loadDeviceWithDeviceId:keyVerificationStart.fromDevice andUserId:event.sender success:^(MXDeviceInfo *otherDevice) {
 
-        MXDeviceVerificationTransaction *existingTransaction = [self transactionWithUser:event.sender andDevice:keyVerificationStart.fromDevice];
+        MXKeyVerificationTransaction *existingTransaction = [self transactionWithUser:event.sender andDevice:keyVerificationStart.fromDevice];
         if (existingTransaction)
         {
             NSLog(@"[MXKeyVerification] handleStartEvent: already existing transaction. Cancel both");
@@ -608,7 +608,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
         }
 
         // Multiple keyshares between two devices: any two devices may only have at most one key verification in flight at a time.
-        NSArray<MXDeviceVerificationTransaction*> *transactionsWithUser = [self transactionsWithUser:event.sender];
+        NSArray<MXKeyVerificationTransaction*> *transactionsWithUser = [self transactionsWithUser:event.sender];
         if (transactionsWithUser.count)
         {
             NSLog(@"[MXKeyVerification] handleStartEvent: already existing transaction with the user. Cancel both");
@@ -660,7 +660,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
     if (cancelContent)
     {
-        MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:cancelContent.transactionId];
+        MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:cancelContent.transactionId];
         if (transaction)
         {
             [transaction handleCancel:cancelContent];
@@ -689,7 +689,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
     if (acceptContent)
     {
-        MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:acceptContent.transactionId];
+        MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:acceptContent.transactionId];
         if (transaction)
         {
             [transaction handleAccept:acceptContent];
@@ -714,7 +714,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
     if (keyContent)
     {
-        MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:keyContent.transactionId];
+        MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:keyContent.transactionId];
         if (transaction)
         {
             [transaction handleKey:keyContent];
@@ -739,7 +739,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
     if (macContent)
     {
-        MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:macContent.transactionId];
+        MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:macContent.transactionId];
         if (transaction)
         {
             [transaction handleMac:macContent];
@@ -788,7 +788,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
 - (void)setupIncomingDMEvents
 {
-    [_crypto.mxSession listenToEventsOfTypes:kMXDeviceVerificationManagerDMEventTypes onEvent:^(MXEvent *event, MXTimelineDirection direction, id customObject) {
+    [_crypto.mxSession listenToEventsOfTypes:kMXKeyVerificationManagerDMEventTypes onEvent:^(MXEvent *event, MXTimelineDirection direction, id customObject) {
         if (direction == MXTimelineDirectionForwards
             && ![event.sender isEqualToString:self.crypto.mxSession.myUser.userId])
         {
@@ -799,7 +799,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
 - (BOOL)isVerificationByDMEventType:(MXEventTypeString)type
 {
-    return [kMXDeviceVerificationManagerDMEventTypes containsObject:type];
+    return [kMXKeyVerificationManagerDMEventTypes containsObject:type];
 }
 
 - (MXHTTPOperation*)sendMessage:(NSString*)userId
@@ -896,8 +896,8 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
             }
             else
             {
-                NSError *error = [NSError errorWithDomain:MXDeviceVerificationErrorDomain
-                                                     code:MXDeviceVerificationUnknownDeviceCode
+                NSError *error = [NSError errorWithDomain:MXKeyVerificationErrorDomain
+                                                     code:MXKeyVerificationUnknownDeviceCode
                                                  userInfo:@{
                                                             NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unknown device: %@:%@", userId, deviceId]
                                                             }];
@@ -923,8 +923,8 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     MXRoom *room = [_crypto.mxSession roomWithRoomId:roomId];
     if (!room)
     {
-        NSError *error = [NSError errorWithDomain:MXDeviceVerificationErrorDomain
-                                             code:MXDeviceVerificationUnknownRoomCode
+        NSError *error = [NSError errorWithDomain:MXKeyVerificationErrorDomain
+                                             code:MXKeyVerificationUnknownRoomCode
                                          userInfo:@{
                                                     NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unknown room: %@", roomId]
                                                     }];
@@ -984,9 +984,9 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
         if (notify)
         {
-            [[NSNotificationCenter defaultCenter] postNotificationName:MXDeviceVerificationManagerNewRequestNotification object:self userInfo:
+            [[NSNotificationCenter defaultCenter] postNotificationName:MXKeyVerificationManagerNewRequestNotification object:self userInfo:
              @{
-               MXDeviceVerificationManagerNotificationRequestKey: request
+               MXKeyVerificationManagerNotificationRequestKey: request
                }];
         }
     }
@@ -1079,20 +1079,20 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
 #pragma mark - Transactions queue
 
-- (MXDeviceVerificationTransaction*)transactionWithUser:(NSString*)userId andDevice:(NSString*)deviceId
+- (MXKeyVerificationTransaction*)transactionWithUser:(NSString*)userId andDevice:(NSString*)deviceId
 {
     return [transactions objectForDevice:deviceId forUser:userId];
 }
 
-- (NSArray<MXDeviceVerificationTransaction*>*)transactionsWithUser:(NSString*)userId
+- (NSArray<MXKeyVerificationTransaction*>*)transactionsWithUser:(NSString*)userId
 {
     return [transactions objectsForUser:userId];
 }
 
-- (MXDeviceVerificationTransaction*)transactionWithTransactionId:(NSString*)transactionId
+- (MXKeyVerificationTransaction*)transactionWithTransactionId:(NSString*)transactionId
 {
-    MXDeviceVerificationTransaction *transaction;
-    for (MXDeviceVerificationTransaction *t in transactions.allObjects)
+    MXKeyVerificationTransaction *transaction;
+    for (MXKeyVerificationTransaction *t in transactions.allObjects)
     {
         if ([t.transactionId isEqualToString:transactionId])
         {
@@ -1104,22 +1104,22 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     return transaction;
 }
 
-- (void)addTransaction:(MXDeviceVerificationTransaction*)transaction
+- (void)addTransaction:(MXKeyVerificationTransaction*)transaction
 {
     [transactions setObject:transaction forUser:transaction.otherUserId andDevice:transaction.otherDeviceId];
     [self scheduleTransactionTimeoutTimer];
 
     dispatch_async(dispatch_get_main_queue(),^{
-        [[NSNotificationCenter defaultCenter] postNotificationName:MXDeviceVerificationManagerNewTransactionNotification object:self userInfo:
+        [[NSNotificationCenter defaultCenter] postNotificationName:MXKeyVerificationManagerNewTransactionNotification object:self userInfo:
          @{
-           MXDeviceVerificationManagerNotificationTransactionKey: transaction
+           MXKeyVerificationManagerNotificationTransactionKey: transaction
            }];
     });
 }
 
 - (void)removeTransactionWithTransactionId:(NSString*)transactionId
 {
-    MXDeviceVerificationTransaction *transaction = [self transactionWithTransactionId:transactionId];
+    MXKeyVerificationTransaction *transaction = [self transactionWithTransactionId:transactionId];
     if (transaction)
     {
         [transactions removeObjectForUser:transaction.otherUserId andDevice:transaction.otherDeviceId];
@@ -1130,7 +1130,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 - (nullable NSDate*)oldestTransactionCreationDate
 {
     NSDate *oldestCreationDate;
-    for (MXDeviceVerificationTransaction *transaction in transactions.allObjects)
+    for (MXKeyVerificationTransaction *transaction in transactions.allObjects)
     {
         if (!oldestCreationDate
             || transaction.creationDate.timeIntervalSince1970 < oldestCreationDate.timeIntervalSince1970)
@@ -1141,7 +1141,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
     return oldestCreationDate;
 }
 
-- (BOOL)isCreationDateValid:(MXDeviceVerificationTransaction*)transaction
+- (BOOL)isCreationDateValid:(MXKeyVerificationTransaction*)transaction
 {
     return (transaction.creationDate.timeIntervalSinceNow > -MXTransactionTimeout);
 }
@@ -1205,7 +1205,7 @@ static NSArray<MXEventTypeString> *kMXDeviceVerificationManagerDMEventTypes;
 
 - (void)checkTransactionTimeouts
 {
-    for (MXDeviceVerificationTransaction *transaction in transactions.allObjects)
+    for (MXKeyVerificationTransaction *transaction in transactions.allObjects)
     {
         if (![self isCreationDateValid:transaction])
         {